Moves SearcBar, ShimmeringResourceLoader, ShimmeringTagListLoader, StorySection
