+++ amundsen_application/static/js/components/Alert/alert.story.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import StorySection from '../common/StorySection';
+import StorySection from '../StorySection';
 import Alert from '.';
 
 const stories = storiesOf('Components/Alert', module);
