+++ amundsen_application/static/js/components/SearchBar/InlineSearchResults/ResultItemList/ResultItem/index.spec.tsx
