+++ amundsen_application/static/js/components/SearchBar/InlineSearchResults/ResultItemList/index.tsx
