+++ amundsen_application/static/js/components/SearchBar/InlineSearchResults/styles.scss
