+++ amundsen_application/static/js/components/SearchBar/styles.scss
