+++ amundsen_application/static/js/components/ShimmeringResourceLoader/index.spec.tsx
