+++ amundsen_application/static/js/components/ShimmeringResourceLoader/styles.scss
