+++ amundsen_application/static/js/features/ColumnList/ColumnStats/columnStats.story.tsx
@@ -4,7 +4,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 
-import StorySection from 'components/common/StorySection';
+import StorySection from 'components/StorySection';
 import ColumnStats from '.';
 
 import TestDataBuilder from './testDataBuilder';
