+++ amundsen_application/static/js/pages/HomePage/index.tsx
@@ -12,7 +12,7 @@ import { UpdateSearchStateReset } from 'ducks/search/types';
 import MyBookmarks from 'components/Bookmark/MyBookmarks';
 import Breadcrumb from 'components/Breadcrumb';
 import PopularTables from 'components/PopularTables';
-import SearchBar from 'components/common/SearchBar';
+import SearchBar from 'components/SearchBar';
 import TagsListContainer from 'components/common/Tags';
 import Announcements from 'components/Announcements';
