+++ amundsen_application/static/js/pages/TableDetailPage/TableDashboardResourceList/index.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 
 import ResourceList from 'components/ResourceList';
-import ShimmeringResourceLoader from 'components/common/ShimmeringResourceLoader';
+import ShimmeringResourceLoader from 'components/ShimmeringResourceLoader';
 
 import { GlobalState } from 'ducks/rootReducer';
